Implement games and rounds
This commit finishes the project. I had a fair amount of help with
this one, but ultimately I was the one writing the code.

I had to ensure that the inputs of the player are correct both
in terms of their case and their letters. I did this using a
while(true) loop. This idea came from my uncle.

I also ensured that the game ran five times, that the score was
displayed each time and that a winner was declared at the end.

I'm not sure if I could recreate this project, but it was a fun
challenge once I had got a little bit of help with it!