Add styles.css, implement buttons
This commit makes it so that there is a styles.css file for later use,
it also adds buttons with event listeners that allow the game
to be played with buttons.

What it has not yet (re)incorporated is the ability to live update
the scores and declare a winner after either the player or computer
has reached a total of 5 wins. This is the next stage of the project.