Add score functionality
This commit adds player choice and score functionality to the
previous iteration of the game. It makes it so that the score is
output and the user can see what the inputs for both parties were.

However! There is a bug in the code which means that the score is
updating a turn after it is supposed to. For this reason I am
committing to the branch here and will try and work on the
if logic.