Correct scoring functions
This commit 'completes' the project by adding and correcting
scoring functions into the mix. I say 'completes' in quotes because
in fact there is still another element to be added (at a future date).
That being a 'play again' feature so that the scores can be reset
and the game can be replayed. Still, for now I am happy with this.